requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cloudinary==1.29.0
Django==4.0.3
djangorestframework==3.13.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
